Load SecurityConfiguration only once
Previously the SecurityConfiguration was loaded in the
ContextLoaderListener's ApplicationContext and in the DispatcherServlet's
ApplicationContext. This meant that there were duplicate objects for the
SecurityConfiguration being created. One specific problem that could occur
is that the InMemoryTokenStore would be created twice and if different
instances were used for storing and retrieving the AccessToken's it would
be impossible to authenticate with OAuth.

Now SecurityConfiguration is only loaded once. To accomodate this, the
SecurityConfiguration is loaded by the DispatcherServlet and we instruct
the DelegatingFilterProxy to load from the Dispatcher's ApplicationContext.
It is necessary to have the Spring Security OAuth configuration created in
the child context in order for Spring MVC to find the additional mappings.