Update sample oauth token in crm-schema-h2.sql
The previously inserted authentication associated to the access token was
using the default JVM generated serialVersionUID for the User. Commit
db03bc2a updated User to have a serialVersionUID of 1L which meant when
OAuth tried to deserialize the User object it failed. This issue manifested
itself as a NullPointerException within the JdbcTokenStore.

I have updated the crm-schema-h2.sql to use a newly generated access token.
This means clients that previously had an association to the access token
will need to update their access token to the new value.