Disable Spring Security for /h2/**
Since h2 has its own security, disable Spring Security for it. This makes
the configuration much simpler and allows re-enabling of CSRF for the rest
of the application. Technically we could just disable CSRF for the h2
portion using http.csrf().requireCsrfProtectionMatcher and leave the
remaining security in place, but this also allows us to remove the headers
customizations.